EMA calculation

+++ src/main/java/com/example/Portfolio.java
@@ -3,9 +3,9 @@ package com.example;
 public class Portfolio {
     float coinAmount;
     float boughtAt;
-    int boughtTime;
+    long boughtTime;
 
-    Portfolio(float boughtAt, float coinAmount, int boughtTime) {
+    Portfolio(float boughtAt, float coinAmount, long boughtTime) {
         this.boughtAt = boughtAt;
         this.coinAmount = coinAmount;
         this.boughtTime = boughtTime;
